Renaming package to homework5.
