+++ src/main/java/com/march/project/homework5/Animals.java
@@ -1,4 +1,4 @@
-package com.march.project.homework4;
+package com.march.project.homework5;
 
 public class Animals {
     String name;
